Split up the MQTT handler into two cleaner files
