+++ Ruby/MQTT/lib/mqtt/base_handler.rb
@@ -1,0 +1,324 @@
+
+require 'timeout'
+require 'mqtt'
+require 'colorize'
+
+module MQTT
+	class BaseHandler
+		# Split a Topic into a Topic-Array
+		# @param topicName [String] The string topic which to split
+		# @return [Array<String>] A list of individual topic-branches
+		# @note This function is mainly used for background processing.
+		def self.get_topic_split(topicName)
+			return topicName.scan(/[^\/]+/);
+		end
+
+		# Match a topic string to a topic pattern
+		# @param receivedTopicString [String] The string (as
+		#  returned by MQTT.get) to compare
+		# @param topicPattern [Array<String>] The Topic-Array (as
+		#  returned by .get_topic_split) to compare against
+		# @return [nil, Array<String>] Nil if no match was found.
+		#  An Array of matched wildcard topic branches (can be empty) when
+		#  successfully matched
+		# @note (see .get_topic_split)
+		def self.getTopicMatch(receivedTopicString, topicPattern)
+			receivedTopicList = get_topic_split receivedTopicString;
+
+			outputTopicList = Array.new();
+
+			return nil unless receivedTopicList.length >= topicPattern.length;
+
+			topicPattern.each_index do |i|
+				if(topicPattern[i] == "+")
+					outputTopicList << receivedTopicList[i];
+
+				elsif(topicPattern[i] == "#")
+					outputTopicList.concat receivedTopicList[i..-1];
+					return outputTopicList;
+
+				elsif topicPattern[i] != receivedTopicList[i];
+					return nil;
+
+				end
+			end
+
+			return outputTopicList if topicPattern.length == receivedTopicList.length;
+			return nil;
+		end
+
+		# Call all existing callbacks whose topic-list matches `topic`
+		def call_interested(topic, data)
+			topicHasReceivers = false;
+			@callbackList.each do |h|
+				tMatch = BaseHandler.getTopicMatch(topic, h.topic_split);
+				if tMatch
+					begin
+						Timeout.timeout(10) {
+							h.offer(tMatch, data)
+						}
+					rescue Timeout::Error
+						STDERR.puts "MQTT Callback Timeout #{h}"
+					end
+					topicHasReceivers = true;
+				end
+			end
+
+			@mqtt.unsubscribe(topic) unless topicHasReceivers;
+		end
+		private :call_interested
+
+		def queue_packet(data)
+			return if @destroying
+
+			@packetQueueMutex.synchronize {
+				@packetQueue << data;
+				@packetQueue.shift if @packetQueue.size > 100
+
+				@publisherThread.run() if @publisherThreadWaiting;
+			}
+		end
+
+		# @!group Custom subscription handling
+
+		# Unregister a subscription. Removes it from the callback list and
+		# unsubscribes from the topic if no other subscriptions for it are present.
+		# @param subObject [MQTT::Subscriptions::Subscription]
+		#  The subscription-object to remove
+		# @return void
+		def unregister_subscription(subObject)
+			raise ArgumentError, "Object is not a subscription!" unless subObject.is_a? MQTT::Subscriptions::Subscription
+			return unless @callbackList.include? subObject;
+
+			queue_packet({type: :unsub, topic: subObject.topic});
+			@callbackList.delete(subObject);
+		end
+		# Register a custom subscription, and send a subscription message to the server.
+		# @param subObject [MQTT::Subscriptions::Subscription]
+		#  An instance of a MQTT Subscription object
+		# @return void
+		def register_subscription(subObject)
+			raise ArgumentError, "Object is not a subscription!" unless subObject.is_a? MQTT::Subscriptions::Subscription
+			return if @callbackList.include? subObject;
+
+			@callbackList << subObject;
+			queue_packet({type: :sub, topic: subObject.topic, qos: subObject.qos});
+		end
+
+		# @!endgroup
+
+		def ensure_clean_start()
+			@mqttWasStartedClean = @mqtt.clean_session
+			if @mqttWasStartedClean
+				begin
+					@mqtt.connect();
+					@mqtt.disconnect();
+				rescue MQTT::Exception
+					sleep 1;
+					retry
+				rescue SocketError, SystemCallError
+					sleep 5
+					retry
+				end
+				@mqtt.clean_session=false;
+			end
+		end
+		private :ensure_clean_start
+
+		def ensure_clean_exit()
+			if(@mqttWasStartedClean)
+				print "Logging out of mqtt server... "
+				begin
+				Timeout.timeout(3) {
+					begin
+						@mqtt.clean_session = true;
+						@mqtt.disconnect();
+						@mqtt.connect();
+					rescue MQTT::Exception, SocketError, SystemCallError
+						sleep 0.3
+						retry;
+					end
+				}
+				rescue  Timeout::Error
+					puts "Timed out, aborting!";
+				else
+					puts "Done."
+				end
+			end
+		end
+		private :ensure_clean_exit
+
+		def mqtt_push_thread
+			loop do
+				@packetQueueMutex.synchronize {
+					@publisherThreadWaiting = true;
+				}
+				Thread.stop();
+				@packetQueueMutex.synchronize {
+					@publisherThreadWaiting = false;
+				}
+				return if @destroying
+
+				next unless @connected
+
+				begin
+					until @packetQueue.empty? do
+						h = nil;
+						@packetQueueMutex.synchronize {
+							h = @packetQueue[0];
+						}
+						Timeout.timeout(3) {
+							if(h[:type] == :sub)
+								@mqtt.subscribe(h[:topic] => h[:qos]);
+							elsif(h[:type] == :pub)
+								@mqtt.publish(h[:topic], h[:data], h[:retain], h[:qos]);
+							end
+						}
+						@packetQueueMutex.synchronize {
+							@packetQueue.shift();
+						}
+					end
+				rescue MQTT::Exception, SocketError, SystemCallError, Timeout::Error => e
+						STDERR.puts("MQTT: #{@mqtt.host} push error, disconnecting!".red) if @connected
+						STDERR.puts(e.inspect);
+
+						sleep 1
+				end
+			end
+		end
+		private :mqtt_push_thread
+
+		def mqtt_resub_thread
+			loop do
+				begin
+					return if @destroying
+
+					STDERR.puts("MQTT: #{@mqtt.host} trying reconnect...".yellow)
+					Timeout.timeout(4) {
+						@mqtt.connect()
+					}
+					STDERR.puts("MQTT: #{@mqtt.host} connected!".green)
+					@conChangeMutex.synchronize {
+						@connected = true;
+					}
+
+					@packetQueueMutex.synchronize {
+						@publisherThread.run() if (@publisherThread && @publisherThreadWaiting)
+					}
+
+					@mqtt.get do |topic, message|
+						call_interested(topic, message);
+					end
+				rescue MQTT::Exception, Timeout::Error, SocketError, SystemCallError
+					STDERR.puts("MQTT: #{@mqtt.host} disconnected!".red) if @connected
+					@connected = false;
+
+					@conChangeMutex.unlock if @conChangeMutex.owned?
+					@mqtt.clean_session = false;
+					sleep 2
+				end
+			end
+		end
+		private :mqtt_resub_thread
+
+		# Pause the main thread and wait for messages.
+		# This is mainly useful when the code has set everything up, but doesn't just want to end.
+		# "INT" is trapped, ensuring a smooth exit on Ctrl-C
+		def lockAndListen()
+			Signal.trap("INT") {
+				exit 0
+			}
+
+			puts "Main thread paused."
+			Thread.stop();
+		end
+
+		def destroy!()
+			return if @destroying
+			@destroying = true;
+
+			unless @packetQueue.empty?
+				print "Finishing sending of MQTT messages ... "
+				begin
+					Timeout.timeout(4) {
+						until @packetQueue.empty? do
+							sleep 0.05;
+						end
+					}
+				rescue Timeout::Error
+					puts "Timed out, aborting."
+				else
+					puts "Done."
+				end
+			end
+
+			@publisherThread.run();
+			@publisherThread.join();
+			@listenerThread.kill();
+
+			@mqtt.disconnect() if @connected
+
+			ensure_clean_exit();
+		end
+
+		# Initialize a new MQTT::SubHandler
+		# The handler immediately connects to the server, and begins receciving and sending.
+		# @param mqttClient [String, MQTT::Client] Either a URI to connect to, or a MQTT::Client
+		#  The URI can be of the form "mqtts://Password@User:URL:port".
+		#  The MQTT client instance can be fully configured, as specified by the MQTT Gem. It must *not* already be connected!
+		# @param jsonify [Boolean] Should Hashes and Arrays input into publish_to be converted to JSON?
+		#  This can be useful to have one less .to_json call. Default is true.
+		# @example Starting the handler
+		#  mqtt = MQTT::SubHandler.new('mqtt.eclipse.org');
+		#  mqtt = MQTT::SubHandler.new(MQTT::Client.new("Your.Client.Opts"))
+		def initialize(mqttClient)
+			@callbackList = Array.new();
+			if mqttClient.is_a? String
+				@mqtt = MQTT::Client.new(mqttClient);
+				@mqtt.clean_session = false;
+			else
+				@mqtt = mqttClient;
+			end
+
+			@conChangeMutex = Mutex.new();
+			@connected 		= false;
+
+			@mqtt.client_id ||= MQTT::Client.generate_client_id("MQTT_Sub_", 8);
+
+			@packetQueue = Array.new();
+			@packetQueueMutex = Mutex.new();
+			puts "PQUEUE IS #{@packetQueueMutex}"
+
+			@publisherThreadWaiting = false;
+
+			@subscribedTopics 	= Hash.new();
+
+			@trackerHash = Hash.new();
+
+			@listenerThread = Thread.new do
+				ensure_clean_start();
+				mqtt_resub_thread();
+			end
+			@listenerThread.abort_on_exception = true;
+
+			begin
+				Timeout.timeout(5) {
+					until(@connected)
+						sleep 0.1;
+					end
+				}
+			rescue Timeout::Error
+				STDERR.puts "MQTT: #{@mqtt.host} did not connect!".red
+			end
+
+			@publisherThread = Thread.new do
+				mqtt_push_thread();
+			end
+			@publisherThread.abort_on_exception = true;
+
+			at_exit {
+				destroy!()
+			}
+		end
+	end
+end
